BetaMoments {actuar}	R Documentation

Raw and Limited Moments of the Beta Distribution

Description

Raw moments and limited moments for the (central) Beta distribution with parameters shape1 and shape2.

Usage

mbeta(order, shape1, shape2)
levbeta(limit, shape1, shape2, order = 1)

Arguments

order
order of the moment.

limit
limit of the loss variable.

shape1, shape2
positive parameters of the Beta distribution.

Details

The kth raw moment of the random variable X is E[X^k] and the kth limited moment at some limit d is E[min(X, d)^k], k > -shape1.

The noncentral beta distribution is not supported.

Value

mbeta gives the kth raw moment and levbeta gives the kth moment of the limited loss variable.

Invalid arguments will result in return value NaN, with a warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

See Also

Beta for details on the beta distribution and functions [dpqr]beta.

Examples

mbeta(2, 3, 4) - mbeta(1, 3, 4)^2
levbeta(10, 3, 4, order = 2)

[Package actuar version 2.3-2 Index]