CTE {actuar}	R Documentation

Conditional Tail Expectation

Description

Conditional Tail Expectation, also called Tail Value-at-Risk.

TVaR is an alias for CTE.

Usage

CTE(x, ...)

## S3 method for class 'aggregateDist'
CTE(x, conf.level = c(0.9, 0.95, 0.99),
         names = TRUE, ...)

TVaR(x, ...)

Arguments

x
an R object.

conf.level
numeric vector of probabilities with values in [0, 1)

.

names
logical; if true, the result has a names attribute. Set to FALSE for speedup with many probs.

...
further arguments passed to or from other methods.

Details

The Conditional Tail Expectation (or Tail Value-at-Risk) measures the average of losses above the Value at Risk for some given confidence level, that is E[X|X > \mathrm{VaR}(X)] where X is the loss random variable.

CTE is a generic function with, currently, only a method for objects of class "aggregateDist".

For the recursive, convolution and simulation methods of aggregateDist, the CTE is computed from the definition using the empirical cdf.

For the normal approximation method, an explicit formula exists:

m + s exp(-VaR(X)^2/2)/((1 - a) * sqrt(2 pi)),

where m is the mean, s the standard deviation and a the confidence level.

For the Normal Power approximation, the explicit formula given in Castañer et al. (2013) is

m + s exp(-VaR(X)^2/2)/((1 - a) * sqrt(2 pi)) (1 + g * VaR(X)/6),

where, as above, m is the mean, s the standard deviation, a the confidence level and g is the skewness.

the CTE is computed from the definition using integrate.

Value

A numeric vector, named if names is TRUE.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Tommy Ouellet

References

Castañer, A. and Claramunt, M.M. and Mármol, M. (2013), Tail value at risk. An analysis with the Normal-Power approximation. In Statistical and Soft Computing Approaches in Insurance Problems, pp. 87-112. Nova Science Publishers, 2013. ISBN 978-1-62618-506-7. https://www.novapublishers.com/catalog/product_info.php?products_id=43105

See Also

aggregateDist; VaR

Examples

model.freq <- expression(data = rpois(7))
model.sev <- expression(data = rnorm(9, 2))
Fs <- aggregateDist("simulation", model.freq, model.sev, nb.simul = 1000)
CTE(Fs)

[Package actuar version 2.2-0 Index]