ExponentialSupp {actuar}	R Documentation

Moments and Moment Generating Function of the Exponential Distribution

Description

Raw moments, limited moments and moment generating function for the exponential distribution with rate rate (i.e., mean 1/rate).

Usage

mexp(order, rate = 1)
levexp(limit, rate = 1, order = 1)
mgfexp(t, rate = 1, log = FALSE)

Arguments

order
order of the moment.

limit
limit of the loss variable.

rate
vector of rates.

t
numeric vector.

log
logical; if TRUE, the cumulant generating function is returned.

Details

The kth raw moment of the random variable X is E[X^k], the kth limited moment at some limit d is E[min(X, d)^k] and the moment generating function is E[e^{tX}], k > -1.

Value

mexp gives the kth raw moment, levexp gives the kth moment of the limited loss variable, and mgfexp gives the moment generating function in t.

Invalid arguments will result in return value NaN, with a warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca, Christophe Dutang and Mathieu Pigeon.

References

Johnson, N. L. and Kotz, S. (1970), Continuous Univariate Distributions, Volume 1, Wiley.

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

See Also

Exponential

Examples

mexp(2, 3) - mexp(1, 3)^2
levexp(10, 3, order = 2)
mgfexp(1,2)

[Package actuar version 2.3-1 Index]