GeneralizedPareto {actuar}	R Documentation

The Generalized Pareto Distribution

Description

Density function, distribution function, quantile function, random generation, raw moments and limited moments for the Generalized Pareto distribution with parameters shape1, shape2 and scale.

Usage

dgenpareto(x, shape1, shape2, rate = 1, scale = 1/rate,
           log = FALSE)
pgenpareto(q, shape1, shape2, rate = 1, scale = 1/rate,
           lower.tail = TRUE, log.p = FALSE)
qgenpareto(p, shape1, shape2, rate = 1, scale = 1/rate,
           lower.tail = TRUE, log.p = FALSE)
rgenpareto(n, shape1, shape2, rate = 1, scale = 1/rate)
mgenpareto(order, shape1, shape2, rate = 1, scale = 1/rate)
levgenpareto(limit, shape1, shape2, rate = 1, scale = 1/rate,
             order = 1)

Arguments

x, q
vector of quantiles.

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

shape1, shape2, scale
parameters. Must be strictly positive.

rate
an alternative way to specify the scale.

log, log.p
logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

order
order of the moment.

limit
limit of the loss variable.

Details

The Generalized Pareto distribution with parameters shape1 = a, shape2 = b and scale = s has density:

f(x) = Gamma(a + b)/(Gamma(a) * Gamma(b)) * (s^a x^(b - 1))/(x + s)^(a + b)

for x > 0, a > 0, b > 0 and s > 0. (Here Γ(α) is the function implemented by R's gamma() and defined in its help.)

The Generalized Pareto is the distribution of the random variable

θ (X/(1 - X)),

where X has a beta distribution with parameters α and τ.

The Generalized Pareto distribution has the following special cases:

A Pareto distribution when shape2 == 1;

An Inverse Pareto distribution when shape1 == 1.

The kth raw moment of the random variable X is E[X^k], -shape2 < k < shape1.

The kth limited moment at some limit d is E[min(X, d)^k], k > -shape2 and shape1 - k not a negative integer.

Value

dgenpareto gives the density, pgenpareto gives the distribution function, qgenpareto gives the quantile function, rgenpareto generates random deviates, mgenpareto gives the kth raw moment, and levgenpareto gives the kth moment of the limited loss variable.

Invalid arguments will result in return value NaN, with a warning.

Note

levgenpareto computes the limited expected value using betaint.

Distribution also known as the Beta of the Second Kind. See also Kleiber and Kotz (2003) for alternative names and parametrizations.

The Generalized Pareto distribution defined here is different from the one in Embrechts et al. (1997) and in Wikipedia; see also Kleiber and Kotz (2003, section 3.12). One may most likely compute quantities for the latter using functions for the Pareto distribution with the appropriate change of parametrization.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Embrechts, P., Klüppelberg, C. and Mikisch, T. (1997), Modelling Extremal Events for Insurance and Finance, Springer.

Kleiber, C. and Kotz, S. (2003), Statistical Size Distributions in Economics and Actuarial Sciences, Wiley.

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

Examples

exp(dgenpareto(3, 3, 4, 4, log = TRUE))
p <- (1:10)/10
pgenpareto(qgenpareto(p, 3, 3, 1), 3, 3, 1)
qgenpareto(.3, 3, 4, 4, lower.tail = FALSE)

## variance
mgenpareto(2, 3, 2, 1) - mgenpareto(1, 3, 2, 1)^2

## case with shape1 - order > 0
levgenpareto(10, 3, 3, scale = 1, order = 2)

## case with shape1 - order < 0
levgenpareto(10, 1.5, 3, scale = 1, order = 2)

[Package actuar version 2.3-2 Index]