InverseTransformedGamma {actuar}	R Documentation

The Inverse Transformed Gamma Distribution

Description

Density function, distribution function, quantile function, random generation, raw moments, and limited moments for the Inverse Transformed Gamma distribution with parameters shape1, shape2 and scale.

Usage

dinvtrgamma(x, shape1, shape2, rate = 1, scale = 1/rate,
            log = FALSE)
pinvtrgamma(q, shape1, shape2, rate = 1, scale = 1/rate,
            lower.tail = TRUE, log.p = FALSE)
qinvtrgamma(p, shape1, shape2, rate = 1, scale = 1/rate,
            lower.tail = TRUE, log.p = FALSE)
rinvtrgamma(n, shape1, shape2, rate = 1, scale = 1/rate)
minvtrgamma(order, shape1, shape2, rate = 1, scale = 1/rate)
levinvtrgamma(limit, shape1, shape2, rate = 1, scale = 1/rate,
              order = 1)

Arguments

x, q
vector of quantiles.

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

shape1, shape2, scale
parameters. Must be strictly positive.

rate
an alternative way to specify the scale.

log, log.p
logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

order
order of the moment.

limit
limit of the loss variable.

Details

The inverse transformed gamma distribution with parameters shape1 = a, shape2 = b and scale = s, has density:

f(x) = b u^a exp(-u) / (x Gamma(a)), u = (s/x)^b

for x > 0, a > 0, b > 0 and s > 0. (Here Gamma(a) is the function implemented by R's gamma() and defined in its help.)

The inverse transformed gamma is the distribution of the random variable s X^(-1/b), where X has a gamma distribution with shape parameter a and scale parameter 1 or, equivalently, of the random variable Y^(-1/b) with Y a gamma distribution with shape parameter a and scale parameter s^(-b).

The inverse transformed gamma distribution defines a family of distributions with the following special cases:

An Inverse Gamma distribution when shape2 == 1;

An Inverse Weibull distribution when shape1 == 1;

An Inverse Exponential distribution when shape1 == shape2 == 1;

The kth raw moment of the random variable X is E[X^k], k < shape1 * shape2, and the kth limited moment at some limit d is E[min(X, d)^k] for all k.

Value

dinvtrgamma gives the density, pinvtrgamma gives the distribution function, qinvtrgamma gives the quantile function, rinvtrgamma generates random deviates, minvtrgamma gives the kth raw moment, and levinvtrgamma gives the kth moment of the limited loss variable.

Invalid arguments will result in return value NaN, with a warning.

Note

levinvtrgamma computes the limited expected value using gammainc from package expint.

Distribution also known as the Inverse Generalized Gamma. See also Kleiber and Kotz (2003) for alternative names and parametrizations.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Kleiber, C. and Kotz, S. (2003), Statistical Size Distributions in Economics and Actuarial Sciences, Wiley.

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

Examples

exp(dinvtrgamma(2, 3, 4, 5, log = TRUE))
p <- (1:10)/10
pinvtrgamma(qinvtrgamma(p, 2, 3, 4), 2, 3, 4)
minvtrgamma(2, 3, 4, 5)
levinvtrgamma(200, 3, 4, 5, order = 2)

[Package actuar version 2.3-1 Index]