LognormalMoments {actuar}	R Documentation

Raw and Limited Moments of the Lognormal Distribution

Description

Raw moments and limited moments for the Lognormal distribution whose logarithm has mean equal to meanlog and standard deviation equal to sdlog.

Usage

mlnorm(order, meanlog = 0, sdlog = 1)
levlnorm(limit, meanlog = 0, sdlog = 1, order = 1)

Arguments

order
order of the moment.

limit
limit of the loss variable.

meanlog, sdlog
mean and standard deviation of the distribution on the log scale with default values of 0 and 1 respectively.

Value

mlnorm gives the kth raw moment and levlnorm gives the kth moment of the limited loss variable.

Invalid arguments will result in return value NaN, with a warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

See Also

Lognormal for details on the lognormal distribution and functions [dpqr]lnorm.

Examples

mlnorm(2, 3, 4) - mlnorm(1, 3, 4)^2
levlnorm(10, 3, 4, order = 2)

[Package actuar version 2.2-0 Index]