NormalSupp {actuar}	R Documentation

Moments and Moment generating function of the Normal Distribution

Description

Raw moments and moment generating function for the normal distribution with mean equal to mean and standard deviation equal to sd.

Usage

mnorm(order, mean = 0, sd = 1)
mgfnorm(t, mean = 0, sd = 1, log = FALSE)

Arguments

order
vector of integers; order of the moment.

mean
vector of means.

sd
vector of standard deviations.

t
numeric vector.

log
logical; if TRUE, the cumulant generating function is returned.

Details

The kth raw moment of the random variable X is E[X^k] and the moment generating function is E[e^{tX}].

Only integer moments are supported.

Value

mnorm gives the kth raw moment and mgfnorm gives the moment generating function in t.

Invalid arguments will result in return value NaN, with a warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca, Christophe Dutang

References

Johnson, N. L. and Kotz, S. (1970), Continuous Univariate Distributions, Volume 1, Wiley.

See Also

Normal

Examples

mgfnorm(0:4,1,2)
mnorm(3)

[Package actuar version 2.3-2 Index]