Pareto {actuar}	R Documentation

The Pareto Distribution

Description

Density function, distribution function, quantile function, random generation, raw moments and limited moments for the Pareto distribution with parameters shape and scale.

Usage

dpareto(x, shape, scale, log = FALSE)
ppareto(q, shape, scale, lower.tail = TRUE, log.p = FALSE)
qpareto(p, shape, scale, lower.tail = TRUE, log.p = FALSE)
rpareto(n, shape, scale)
mpareto(order, shape, scale)
levpareto(limit, shape, scale, order = 1)

Arguments

x, q
vector of quantiles.

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

shape, scale
parameters. Must be strictly positive.

log, log.p
logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

order
order of the moment.

limit
limit of the loss variable.

Details

The Pareto distribution with parameters shape = a and scale = s has density:

f(x) = a s^a / (x + s)^(a + 1)

for x > 0, a > 0 and s > 0.

The kth raw moment of the random variable X is E[X^k], -1 < k < shape.

The kth limited moment at some limit d is E[min(X, d)^k], k > -1 and shape - k not a negative integer.

Value

dpareto gives the density, ppareto gives the distribution function, qpareto gives the quantile function, rpareto generates random deviates, mpareto gives the kth raw moment, and levpareto gives the kth moment of the limited loss variable.

Invalid arguments will result in return value NaN, with a warning.

Note

levpareto computes the limited expected value using betaint.

Distribution also known as the Pareto Type II or Lomax distribution. See also Kleiber and Kotz (2003) for alternative names and parametrizations.

The version of the Pareto defined for x > s is named Single Parameter Pareto in actuar.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Kleiber, C. and Kotz, S. (2003), Statistical Size Distributions in Economics and Actuarial Sciences, Wiley.

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

See Also

dpareto1 for the Single Parameter Pareto distribution.

Examples

exp(dpareto(2, 3, 4, log = TRUE))
p <- (1:10)/10
ppareto(qpareto(p, 2, 3), 2, 3)

## variance
mpareto(2, 4, 1) - mpareto(1, 4, 1)^2

## case with shape - order > 0
levpareto(10, 3, scale = 1, order = 2)

## case with shape - order < 0
levpareto(10, 1.5, scale = 1, order = 2)

[Package actuar version 2.3-3 Index]