WeibullMoments {actuar}	R Documentation

Raw and Limited Moments of the Weibull Distribution

Description

Raw moments and limited moments for the Weibull distribution with parameters shape and scale.

Usage

mweibull(order, shape, scale = 1)
levweibull(limit, shape, scale = 1, order = 1)

Arguments

order
order of the moment.

limit
limit of the loss variable.

shape, scale
shape and scale parameters, the latter defaulting to 1.

Details

The kth raw moment of the random variable X is E[X^k] and the kth limited moment at some limit d is E[min(X, d)^k], k > -shape.

Value

mweibull gives the kth raw moment and levweibull gives the kth moment of the limited loss variable.

Invalid arguments will result in return value NaN, with a warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

See Also

Weibull for details on the Weibull distribution and functions [dpqr]weibull.

Examples

mweibull(2, 3, 4) - mweibull(1, 3, 4)^2
levweibull(10, 3, 4, order = 2)

[Package actuar version 2.2-0 Index]