aggregate.simpf(actuar)	R Documentation

Compute Summary Statistics of a Portfolio

Description

aggregate method for class "simpf". Splits portfolio data into subsets, computes summary statistics for each.

frequency method for class "simpf". Computes the frequency of claims for subsets of portfolio data.

Usage


## S3 method for class 'simpf':
aggregate(x, by = c("contract", "year"),
          FUN = sum, ...)

## S3 method for class 'simpf':
frequency(x, by = c("contract", "year"), ...)

Arguments

x 	 a simpf object, typically created by simpf

by 	 vector, indicates how data should be aggregated

FUN 	 the function to be applied to data subsets

... 	 potential further arguments required by generic.

Details

aggregate.simpf and frequency.simpf are respectively methods for functions aggregate and frequency, both from package stats. Both methods offer the possibility to aggregate the claims with respect to the contracts, the years, both or neither through argument by. Using frequency.simpf is equivalent to using aggregate.simpf with argument FUN equal to length.

Value

The object returned depends on the way data is aggregated with respect to the call:
If data is aggregated by both contract and year, a matrix will be returned with FUN applied to all losses vectors in all years and in all contracts. Therefore, the dimensions of this resulting matrix will be the same as the dimensions of the data matrix listed in the simpf object, as long as order is respected in the ‘by’ vector. Otherwise, the resulting matrix will be transposed.
If the argument to by is of length 1, function FUN will be applied to either each contract for all years or to each year for all contracts. Therefore, the resulting object will be a vector of length equal to either the number of years or the number of contracts.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Louis-Philippe Pouliot

See Also

aggregate simpf severity frequency frequency.simpf

Examples


modelfreq <- list(dist1 = "pois",
                  par1 = list(lambda = quote(Lambda * weights)),
                  dist2 = "gamma",
                  par2 = c(shape = 2, rate = 1))
modelsev<-list(dist1 = "lnorm",
               par1 = list(meanlog = quote(Theta), sdlog = 1),
               dist2 = "norm",
               par2 = c(mean = 5, sd = 1))
data(hachemeister)
weights <- hachemeister$weights/mean(hachemeister$weights)
x <- simpf(5, 12, modelfreq, modelsev, weights)
aggregate(x, by = "contract", FUN = mean)
frequency(x)

[Package actuar version 0.1-5 Index]