bstraub(actuar)	R Documentation

Buhlmann-Straub Credibility Model

Description

bstraub calculates credibility premiums in the Bühlmann-Straub credibility model, creating an object of class bstraub.

Usage


bstraub(ratios, weights,
        heterogeneity = c("iterative", "unbiased"),
        TOL = 1e-06, echo = FALSE)
## S3 method for class 'bstraub':
print(x,...)

Arguments

ratios 	 matrix of ratios (contracts in lines, years in columns)

weights 	 matrix of weights corresponding to ratios

heterogeneity 	 estimator of the between contract heterogeneity parameter used in premium calculation; "iterative" for the Bischel-Straub estimator; "unbiased" for the usual Bühlmann-Straub estimator (see below)

TOL 	 maximum relative error in the iterative procedure

echo 	 boolean, whether to echo iterative procedure or not

x 	 a bstraub object.

... 	 potential further arguments required by generic.

Details

The credibility premium of contract i is given by

z[i] X[iw] + (1 - z[i]) X[zw],

where

z[i] = (w[i.] a)/(w[i] a + s^2),

X[iw] is the weighted average of the ratios of contract i, X[zw] is the weighted average of the matrix of ratios using credibility factors and w[i.] is the total weight of a contract. s^2 is the estimator of the within contract heterogeneity and a is the estimator of the between contract heterogeneity.

Missing data are represent by NA in both the matrix of ratios and the matrix of weights. The function can cope with complete lines of NA in case a contract has no experience.

Value

Returns an object of class "bstraub". Function summary can be used to display further information on the premiums, such as the values of the structure parameters.
An object of class "bstraub" is a list with the following components:

model 	 the name of the model used ("Bühlmann" or "Bühlmann-Straub")

premiums 	 vector of credibility premiums

individual 	 vector of contract weighted averages

collective 	 collective premium estimator

weights 	 vector of contracts total weights, as used in credibility factors

s2 	 estimator of the within contract heterogeneity parameter

unbiased 	 unbiased estimator of the between contract heterogeneity parameter

iterative 	 iterative estimator of the between contract heterogeneity parameter

Estimation of a

The Bühlmann-Straub unbiaised estimator (heterogeneity = "unbiased") of the between contracts heterogeneity parameter is

a = c sum(w[i.] * (X[iw] - X[ww])^2 - (I - 1) * s^2),

where c = w[..]/(w[..]^2 - sum(w[i.]^2)) and I is the number of contracts.

The Bishel-Straub pseudo-estimator (heterogeneity = "iterative") is obtained recursively as the solution of

a = 1/(I - 1) sum(z[i] * (X[iw] - X[zw])^2).

The fixed point algorithm is used up, with a relative error of TOL stopping criteria.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Sébastien Auclair

References

Goulet, V. (1998), Principles and Application of Credibility Theory, Journal of Actuarial Practice, Volume 6, ISSN 1064-6647.

Goovaerts, M. J. and Kaas, R. and van Heerwaarden, A. E. and Bauwelinckx, T. (1990), Effective actuarial methods, North-Holland.

Examples


data(hachemeister)

## Credibility premiums calculated with the iterative estimator
bstraub(hachemeister$claims, hachemeister$weights)

## Credibility premiums calculated with the unbiased estimator
(x <- bstraub(hachemeister$claims, hachemeister$weights, heterogeneity = "unbiased"))
summary(x)

[Package actuar version 0.1-5 Index]