Burr(actuar)	R Documentation

The Burr Distribution

Description

Density, distribution function, quantile function and random generation for the Burr distribution with parameters shape1, scale and shape2.

Usage


  dburr(x, shape1, scale, shape2, log = FALSE)
  pburr(q, shape1, scale, shape2, lower.tail = TRUE,
        log.p = FALSE)
  qburr(p, shape1, scale, shape2, lower.tail = TRUE,
        log.p = FALSE)
  rburr(n, shape1, scale, shape2)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape1, scale, shape2 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Burr distribution with parameters shape1 = a, scale = b and shape2 = c has density:

p(x) = (a c (x / b)^c) / (x [1 + (x / b)^c]^(a + 1))

for x > 0, a > 0, b > 0 and c > 0.

The Burr distribution has the following special cases:

shape2 == 1 is a Pareto distribution,
shape2 == shape1 is a Paralogistic distribution, and
shape1 == 1 is a Loglogistic distribution.

Value

dburr gives the density, pburr gives the distribution function, qburr gives the quantile function, and rburr generates random deviates.
Invalid arguments will return warning.

Note

Also known as Burr Type XII and Singh-Maddala.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dburr(2, 3, 4, 5, log = TRUE))
p <- (1:10)/10
pburr(qburr(p, 2, 3, 1), 2, 3, 1)

[Package actuar version 0.1-5 Index]