emm {actuar}	R Documentation

Empirical Moments

Description

Raw empirical moments for individual and grouped data.

Usage

emm(x, order = 1, ...)

## Default S3 method:
emm(x, order = 1, ...)

## S3 method for class 'grouped.data'
emm(x, order = 1, ...)

Arguments

x
a vector or matrix of individual data, or an object of class "grouped data".

order
order of the moment. Must be positive.

...
further arguments passed to or from other methods.

Details

Arguments ... are passed to colMeans; na.rm = TRUE may be useful for individual data with missing values.

For individual data, the kth empirical moment is sum(j; x[j]^k).

For grouped data with group boundaries c[1], …, c[r] and group frequencies n[1], …, n[r], the kth empirical moment is

sum(j; (n[j]*(c[j]^k - c[j-1]^k)/(n*(k+1)*(c[j] - c[j-1])))),

where n = sum(j; n[j]).

Value

A named vector or matrix of moments.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (1998), Loss Models, From Data to Decisions, Wiley.

See Also

mean and mean.grouped.data for simpler access to the first moment.

Examples

## Individual data
data(dental)
emm(dental, order = 1:3)

## Grouped data
data(gdental)
emm(gdental)
x <- grouped.data(cj = gdental[, 1],
                  nj1 = sample(1:100, nrow(gdental)),
                  nj2 = sample(1:100, nrow(gdental)))
emm(x) # same as mean(x)

[Package actuar version 2.3-3 Index]