Genpareto(actuar)	R Documentation

The Generalized Pareto Distribution

Description

Density, distribution function, quantile function and random generation for the Generalized Pareto distribution with parameters shape1, scale and shape2.

Usage


  dgenpareto(x, shape1, scale, shape2, log = FALSE)
  pgenpareto(q, shape1, scale, shape2, lower.tail = TRUE,
             log.p = FALSE)
  qgenpareto(p, shape1, scale, shape2, lower.tail = TRUE,
             log.p = FALSE)
  rgenpareto(n, shape1, scale, shape2)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape1, scale, shape2 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Generalized Pareto distribution with parameters shape1 = a, scale = b and shape2 = c has density:

p(x) = (gamma(a + c) / gamma(a)*gamma(c))((b^a x^(b-1)) / (x + b)^(a + c))

for x > 0, a > 0, b > 0 and c > 0. Here Gamma(a) is the function implemented by R's gamma() and defined in its help.

Let Y be a Beta random variable with parameters shape1 and shape2. Then,

scale (Y / (1 - Y))

has the Generalized Pareto probability density function with parameters shape1, scale and shape2.

The Generalized Pareto distribution has the following special cases:

shape2 == 1 is a Pareto distribution, and
shape2 == shape1 == 1 is an Inverse Pareto distribution.

Value

dgenpareto gives the density, pgenpareto gives the distribution function, qgenpareto gives the quantile function, and rgenpareto generates random deviates.
Invalid arguments will return warning.

Note

Also known as the Beta of the second kind.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dgenpareto(3, 3, 4, 4, log = TRUE))
p <- (1:10)/10
pgenpareto(qgenpareto(p, 3, 3, 1), 3, 3, 1)
qgenpareto(.3, 3, 4, 4, lower.tail = FALSE)

[Package actuar version 0.1-5 Index]