hachemeister {actuar}	R Documentation

Hachemeister Data Set

Description

Hachemeister (1975) data set giving average claim amounts in private passenger bodily injury insurance in five U.S. states over 12 quarters between July 1970 and June 1973 and the corresponding number of claims.

Usage

hachemeister

Format

A matrix with 5 rows and the following 25 columns:

state

the state number;

ratio.1, ..., ratio.12

the average claim amounts;

weight.1, ..., weight.12

the corresponding number of claims.

Source

Hachemeister, C. A. (1975), Credibility for regression models with application to trend, Proceedings of the Berkeley Actuarial Research Conference on Credibility, Academic Press.

[Package actuar version 2.2-0 Index]