Invburr(actuar)	R Documentation

The Inverse Burr Distribution

Description

Density, distribution function, quantile function and random generation for the Inverse Burr distribution with parameters shape1, scale and shape2.

Usage


  dinvburr(x, shape1, scale, shape2, log = FALSE)
  pinvburr(q, shape1, scale, shape2, lower.tail = TRUE,
        log.p = FALSE)
  qinvburr(p, shape1, scale, shape2, lower.tail = TRUE,
        log.p = FALSE)
  rinvburr(n, shape1, scale, shape2)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape1, scale, shape2 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Inverse Burr distribution with parameters shape1 = a, scale = b and shape2 = c has density:

p(x) = a c (x/b)^(a c) / (x (1 + (x/b)^c)^(a + 1))

for x > 0, a > 0, b > 0 and c > 0.

The Inverse Burr distribution has the following special cases:

shape2 == 1 is an Inverse Pareto distribution,
shape2 == shape1 is an Inverse Paralogistic distribution, and
shape1 == 1 is a Loglogistic distribution.

Value

dinvburr gives the density, invburr gives the distribution function, qinvburr gives the quantile function, and rinvburr generates random deviates.
Invalid arguments will return warning.

Note

Also known as Dagum distribution.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

See Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dinvburr(2, 3, 4, 5, log = TRUE))
p <- (1:10)/10
pinvburr(qinvburr(p, 2, 3, 1), 2, 3, 1)

[Package actuar version 0.1-5 Index]