Invexp(actuar)	R Documentation

The Inverse Exponential Distribution

Description

Density, distribution function, quantile function and random generation for the Inverse Exponential distribution with parameter scale.

Usage


  dinvexp(x, scale, log = FALSE)
  pinvexp(q, scale, lower.tail = TRUE, log.p = FALSE)
  qinvexp(p, scale, lower.tail = TRUE, log.p = FALSE)
  rinvexp(n, scale)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

scale 	 parameter. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Inverse Exponential distribution with parameter scale = a has density:

p(x) = a exp(-a / x) / x^2

for x > 0 and a > 0.

Value

dinvexp gives the density, pinvexp gives the distribution function, qinvexp gives the quantile function, and rinvexp generates random deviates.
Invalid arguments will return warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dinvexp(2, 2, log = TRUE))
p <- (1:10)/10
pinvexp(qinvexp(p, 2), 2)

[Package actuar version 0.1-5 Index]