Invgamma(actuar)	R Documentation

The Inverse Gamma Distribution

Description

Density, distribution function, quantile function and random generation for the Inverse Gamma distribution with parameters shape and scale.

Usage


  dinvgamma(x, shape, scale, log = FALSE)
  pinvgamma(q, shape, scale, lower.tail = TRUE,
           log.p = FALSE)
  qinvgamma(p, shape, scale, lower.tail = TRUE,
           log.p = FALSE)
  rinvgamma(n, shape, scale)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape, scale 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Inverse Gamma distribution with parameters shape = a and scale = b has density:

p(x) = (b/x)^a exp(-b/x) / (x gamma(a))

for x > 0, a > 0 and b > 0. Here Gamma(a) is the function implemented by R's gamma() and defined in its help.

The Inverse Gamma distribution has the following special case:

shape1 == 1 is an Inverse Exponential distribution.

Value

dinvgamma gives the density, pinvgamma gives the distribution function, qinvgamma gives the quantile function, and rinvgamma generates random deviates.
Invalid arguments will return warning.

Note

Also known as Vinci distribution.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dinvgamma(2, 3, 4, log = TRUE))
p <- (1:10)/10
pinvgamma(qinvgamma(p, 2, 3), 2, 3)

[Package actuar version 0.1-5 Index]