Invparalogis(actuar)	R Documentation

The Inverse Paralogistic Distribution

Description

Density, distribution function, quantile function and random generation for the Inverse Paralogistic distribution with parameters shape and scale.

Usage


  dinvparalogis(x, shape, scale, log = FALSE)
  pinvparalogis(q, shape, scale, lower.tail = TRUE,
           log.p = FALSE)
  qinvparalogis(p, shape, scale, lower.tail = TRUE,
           log.p = FALSE)
  rinvparalogis(n, shape, scale)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape, scale 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Inverse Paralogistic distribution with parameters shape = a and scale = b has density:

p(x) = a^2 (x / b)^(a^2) / (x(1 + (x/b)^a)^(a + 1))

for x > 0, a > 0 and b > 0.

Value

dinvparalogis gives the density, pinvparalogis gives the distribution function, qinvparalogis gives the quantile function, and rinvparalogis generates random deviates.
Invalid arguments will return warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dinvparalogis(2, 3, 4, log = TRUE))
p <- (1:10)/10
pinvparalogis(qinvparalogis(p, 2, 3), 2, 3)

[Package actuar version 0.1-5 Index]