Invweibull(actuar)	R Documentation

The Inverse Weibull Distribution

Description

Density, distribution function, quantile function and random generation for the Inverse Weibull distribution with parameters scale and shape.

Usage


  dinvweibull(x, scale, shape, log = FALSE)
  pinvweibull(q, scale, shape, lower.tail = TRUE,
           log.p = FALSE)
  qinvweibull(p, scale, shape, lower.tail = TRUE,
           log.p = FALSE)
  rinvweibull(n, scale, shape)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

scale, shape 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Inverse Weibull distribution with parameters scale = a and shape = b has density:

p(x) = b (a/x)^b exp((-a/x)^b) / x

for x > 0, a > 0 and b > 0.

The Inverse Weibull distribution has the following special case:

shape == 1 is an Inverse Exponential distribution.

Value

dinvweibull gives the density, pinvweibull gives the distribution function, qinvweibull gives the quantile function, and rinvweibull generates random deviates.
Invalid arguments will return warning.

Note

Also knonw as log-Gompertz distribution.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dinvweibull(2, 3, 4, log = TRUE))
p <- (1:10)/10
pinvweibull(qinvweibull(p, 2, 3), 2, 3)

[Package actuar version 0.1-5 Index]