Lgamma(actuar)	R Documentation

The Loggamma Distribution

Description

Density, distribution function, quantile function and random generation for the Loggamma distribution with parameters shapelog and ratelog.

Usage


  dlgamma(x, shapelog, ratelog, log = FALSE)
  plgamma(q, shapelog, ratelog, lower.tail = TRUE,
           log.p = FALSE)
  qlgamma(p, shapelog, ratelog, lower.tail = TRUE,
           log.p = FALSE)
  rlgamma(n, shapelog, ratelog)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shapelog, ratelog 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Loggamma distribution with parameters shapelog = a and ratelog = b has density:

dgamma(log(x), shapelog, 1.0 / ratelog)

for x > 1, a > 0 and b > 0.

Let X be a Gamma random variable with parameters shapelog and ratelog. Then,

Y = exp(X)

has the Loggamma probability density function with parameters shapelog and ratelog.

Value

dlgamma gives the density, plgamma gives the distribution function, qlgamma gives the quantile function, and rlgamma generates random deviates.
Invalid arguments will return warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix of Hogg & Klugman, Loss Distributions, Wiley, 1984.

Examples


exp(dlgamma(2, 3, 4, log = TRUE))
p <- (1:10)/10
plgamma(qlgamma(p, 2, 3), 2, 3)

[Package actuar version 0.1-5 Index]