Llogis(actuar)	R Documentation

The Loglogistic Distribution

Description

Density, distribution function, quantile function and random generation for the Loglogistic distribution with parameters shape and scale.

Usage


  dllogis(x, shape, scale, log = FALSE)
  pllogis(q, shape, scale, lower.tail = TRUE, log.p = FALSE)
  qllogis(p, shape, scale, lower.tail = TRUE, log.p = FALSE)
  rllogis(n, shape, scale)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape, scale 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Loglogistic distribution with parameters shape = a and scale = b has density:

p(x) = a (x/b)^a /(x(1 + (x/b)^a)^2)

for x > 0, a > 0 and b > 0.

Value

dllogis gives the density, pllogis gives the distribution function, qllogis gives the quantile function, and rllogis generates random deviates.
Invalid arguments will return warning.

Note

Also known as Fisk distribution.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dllogis(2, 3, 4, log = TRUE))
p <- (1:10)/10
pllogis(qllogis(p, 2, 3), 2, 3)

[Package actuar version 0.1-5 Index]