Paralogis(actuar)	R Documentation

The Paralogistic Distribution

Description

Density, distribution function, quantile function and random generation for the Paralogistic distribution with parameters shape and scale.

Usage


  dparalogis(x, shape, scale, log = FALSE)
  pparalogis(q, shape, scale, lower.tail = TRUE, log.p = FALSE)
  qparalogis(p, shape, scale, lower.tail = TRUE, log.p = FALSE)
  rparalogis(n, shape, scale)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape, scale 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Paralogistic distribution with parameters shape = a and scale = b has density:

p(x) = a^2 (x/b)^a / (x(1 + (x/b)^a)^(a+1))

for x > 0, a > 0 and b > 0.

Value

dparalogis gives the density, pparalogis gives the distribution function, qparalogis gives the quantile function, and rparalogis generates random deviates.
Invalid arguments will return warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dparalogis(2, 3, 4, log = TRUE))
p <- (1:10)/10
pparalogis(qparalogis(p, 2, 3), 2, 3)

[Package actuar version 0.1-5 Index]