Pareto(actuar)	R Documentation

The Pareto Distribution

Description

Density, distribution function, quantile function and random generation for the Pareto distribution with parameters shape and scale.

Usage


  dpareto(x, shape, scale, log = FALSE)
  ppareto(q, shape, scale, lower.tail = TRUE, log.p = FALSE)
  qpareto(p, shape, scale, lower.tail = TRUE, log.p = FALSE)
  rpareto(n, shape, scale)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape, scale 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Pareto distribution with parameters shape = a and scale = b has density:

p(x) = a b^a / (x + b)^(a + 1)

for x > 0, a > 0 and b > 0.

Value

dpareto gives the density, ppareto gives the distribution function, qpareto gives the quantile function, and rpareto generates random deviates.
Invalid arguments will return warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dpareto(2, 3, 4, log = TRUE))
p <- (1:10)/10
ppareto(qpareto(p, 2, 3), 2, 3)

[Package actuar version 0.1-5 Index]