Pareto1(actuar)	R Documentation

The Single-parameter Pareto Distribution

Description

Density, distribution function, quantile function and random generation for the Single-parameter Pareto distribution with parameters shape and min.

Usage


  dpareto1(x, shape, min, log = FALSE)
  ppareto1(q, shape, min, lower.tail = TRUE, log.p = FALSE)
  qpareto1(p, shape, min, lower.tail = TRUE, log.p = FALSE)
  rpareto1(n, shape, min)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape, min 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Single-parameter Pareto distribution with parameters shape = a and min = b has density:

p(x) = a b^a / x^(a + 1)

for x > 0, a > 0 and b > 0.

Although there appears to be two parameters, only shape is a true parameter. The value of min must be set in advance.

Value

dpareto1 gives the density, ppareto1 gives the distribution function, qpareto1 gives the quantile function, and rpareto1 generates random deviates.
Invalid arguments will return warning.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dpareto1(5, 3, 4, log = TRUE))
p <- (1:10)/10
ppareto1(qpareto1(p, 2, 3), 2, 3)

[Package actuar version 0.1-5 Index]