quantile.aggregateDist {actuar}	R Documentation

Quantiles of Aggregate Claim Amount Distribution

Description

Quantile and Value-at-Risk methods for objects of class "aggregateDist".

Usage

## S3 method for class 'aggregateDist'
quantile(x, 
         probs = c(0.25, 0.5, 0.75, 0.9, 0.95, 0.975, 0.99, 0.995),
         smooth = FALSE, names = TRUE, ...)

## S3 method for class 'aggregateDist'
VaR(x, conf.level = c(0.9, 0.95, 0.99),
         smooth = FALSE, names = TRUE, ...)

Arguments

x
an object of class "aggregateDist".

probs, conf.level
numeric vector of probabilities with values in [0, 1).

smooth
logical; when TRUE and x is a step function, quantiles are linearly interpolated between knots.

names
logical; if true, the result has a names attribute. Set to FALSE for speedup with many probs.

...
further arguments passed to or from other methods.

Details

The quantiles are taken directly from the cumulative distribution function defined in x. Linear interpolation is available for step functions.

Value

A numeric vector, named if names is TRUE.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Louis-Philippe Pouliot

See Also

aggregateDist

Examples

model.freq <- expression(data = rpois(3))
model.sev <- expression(data = rlnorm(10, 1.5))
Fs <- aggregateDist("simulation", model.freq, model.sev, nb.simul = 1000)
quantile(Fs, probs = c(0.25, 0.5, 0.75))
VaR(Fs)

[Package actuar version 2.2-0 Index]