rearrangepf(actuar)	R Documentation

Reorganization of a portfolio

Description

rearrangepf reorganizes a portfolio of claim amounts stored in a two dimension list.

Usage


rearrangepf(pf)

Arguments

pf 	 a two dimension list containing individual claim amounts

Details

This function takes individual claim amounts of a portfolio of contracts stored in a two dimension list (where each element of the list is a vector of claim amounts) and returns matrices of total claim amounts, number of claims and individual claim amounts.

Value

A list made of:

aggregate 	 matrix of aggregate losses for each contract and each year

frequencies 	 matrix of the number of claims per contract and per year

severities 	 a list of two elements:

claims

matrix of the individual claim amounts for the first n - 1 years of observations

claims.last

matrix of the individual claim amounts for the last year of observations

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Sébastien Auclair

See Also

simpf to create a portfolio of data

Examples


modelfreq <- list(dist1 = "pois",
                  par1 = list(lambda = quote(Lambda * weights)),
                  dist2 = "gamma",
                  par2 = c(shape = 2, rate = 1))
modelsev<-list(dist1 = "lnorm",
               par1 = list(meanlog = quote(Theta), sdlog = 1),
               dist2 = "norm",
               par2 = c(mean = 5, sd = 1))
( x <- simpf(25, 4, modelfreq, modelsev) )
rearrangepf(x$data)

[Package actuar version 0.1-5 Index]