severity {actuar}	R Documentation

Manipulation of Individual Claim Amounts

Description

severity is a generic function created to manipulate individual claim amounts. The function invokes particular methods which depend on the class of the first argument.

Usage

severity(x, ...)

## Default S3 method:
severity(x, bycol = FALSE, drop = TRUE, ...)

Arguments

x
an R object.

bycol
logical; whether to “unroll” horizontally (FALSE) or vertically (TRUE)

...
further arguments to be passed to or from other methods.

drop
logical; if TRUE, the result is coerced to the lowest possible dimension.

Details

Currently, the default method is equivalent to unroll. This is liable to change since the link between the name and the use of the function is rather weak.

Value

A vector or matrix.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Louis-Philippe Pouliot

See Also

severity.portfolio for the original motivation of these functions.

Examples

x <- list(c(1:3), c(1:8), c(1:4), c(1:3))
(mat <- matrix(x, 2, 2))
severity(mat)
severity(mat, bycol = TRUE)

[Package actuar version 2.3-1 Index]