summary.bstraub(actuar)	R Documentation

Summary Method for BSTRAUB Objects

Description

Summarizes a bstraub object and returns an object of class summary.bstraub. There is a print method for the latter.

Usage


## S3 method for class 'bstraub':
summary(object, ...)

## S3 method for class 'summary.bstraub':
print(x, ...)

Arguments

x, object 	 a bstraub object.

... 	 arguments passed to or from other methods.

Details

This function is a method for the generic function summary() for class bstraub.It can be invoked by calling summary(x) for an object x of the appropriate class, or directly by calling summary.bstraub(x).

Value

Returns an object of class summary.bstraub. A print method is defined to display such objects. It first returns the structure parameters estimated to compute the credibility premium. Four characteristics proper to each contract are subsequently displayed in a table.

Collective premium 	 collective premium estimator

Average contract variance 	 estimator of the within contract heterogeneity parameter

Portefolio heterogeneity 	 both unbiased and iterative estimators of the between contract heterogeneity parameter. The iterative estimator is NULL if argument heterogeneity = "unbiased" is passed to bstraub.

Credibility constant 	 estimator of the credibility constant, given by K = s^2 / a, where s^2 is the average contract variance and a the portfolio heterogeneity.

Ind. premium 	 individual experience weighted average

Weight 	 contract total weight, as used in credibility factors

Cred. factor 	 factor used with its complement in the linear expression of the credibility premium

Cred. premium 	 credibility premium for each contract

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Louis-Philippe Pouliot

Examples


data(hachemeister)

## Credibility premiums calculated with the unbiased estimator
(x <- bstraub(hachemeister$claims,
              hachemeister$weights,
              heterogeneity ="unbiased"))
summary(x)

[Package actuar version 0.1-5 Index]