summary.simpf(actuar)	R Documentation

Summary Method for SIMPF Objects

Description

Summarizes a simpf object by displaying the individual claim amounts for the desired years and contracts. The function returns an object of class summary.simpf. There is a print method for the latter.

Usage


## S3 method for class 'simpf':
summary(object, contracts = seq(length = dim(object$data)[1]),
       years = seq(length = dim(object$data)[2]), ...)

## S3 method for class 'summary.simpf':
print(x, contracts = x$contracts,
       years = x$years, ...)

Arguments

x, object 	 a simpf object.

contracts 	 a vector of contracts for which claim amounts should be displayed. All contracts are displayed by default.

years 	 a vector of years for which claim amounts should be displayed. All years observed are displayed by default

... 	 arguments passed to or from other methods.

Details

This function is a method for the generic function summary() for class simpf.It can be invoked by calling summary(x) for an object x of the appropriate class, or directly by calling summary.simpf(x).

Value

Returns an object of class summary.simpf. A print method is defined to display such objects. The latter displays a list of the chosen contracts (passed on to the argument contracts), in which each element contains itself a list of claim amounts for each year passed on to the argument years.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Louis-Philippe Pouliot

Examples


modelfreq <- list(dist1 = "pois",
                  par1 = list(lambda = quote(Lambda * weights)),
                  dist2 = "gamma",
                  par2 = c(shape = 2, rate = 1))
modelsev<-list(dist1 = "lnorm",
               par1 = list(meanlog = quote(Theta), sdlog = 1),
               dist2 = "norm",
               par2 = c(mean = 5, sd = 1))
data(hachemeister)
weights <- hachemeister$weights/mean(hachemeister$weights)
(x <- simpf(5, 12, model.freq = modelfreq, modelsev))
summary(x, contracts = 1:2, years = 10:12)

[Package actuar version 0.1-5 Index]