Trbeta(actuar)	R Documentation

The Transformed Beta Distribution

Description

Density, distribution function, quantile function and random generation for the Transformed Beta distribution with parameters shape1, scale, shape2 and shape3.

Usage


  dtrbeta(x, shape1, scale, shape2, shape3, log = FALSE)
  ptrbeta(q, shape1, scale, shape2, shape3, lower.tail = TRUE,
          log.p = FALSE)
  qtrbeta(p, shape1, scale, shape2, shape3, lower.tail = TRUE,
          log.p = FALSE)
  rtrbeta(n, shape1, scale, shape2, shape3)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape1, scale, shape2, shape3 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Transformed Beta distribution with parameters shape1 = a, scale = b, shape2 = c and shape3 = d has density:

p(x) = (gamma(a + d) / gamma(a)*gamma(d))((c (x / b)^(a c)) / (x (1 + (x / b)^c)^(a + d)))

for x > 0, a > 0, b > 0, c > 0 and d > 0. Here Gamma(a) is the function implemented by R's gamma() and defined in its help.

Let Y be a Beta random variable with parameters shape1 and shape3. Then,

scale (Y / (1 - Y))^(1 / shape2)

has the Transformed Beta probability density function with parameters shape1, scale, shape2 and shape3.

The Transformed Beta distribution defines a family of distributions with the following special cases:

shape3 == 1 is a Burr distribution,
shape3 == shape2 == 1 is a Pareto distribution,
shape3 == shape1 == 1 is a Loglogistic distribution,
shape3 == 1 and shape2 == shape1 is a Paralogistic distribution,
shape2 == 1 is a Generalized Pareto distribution,
shape2 == shape1 == 1 is an Inverse Pareto distribution,
shape1 == 1 is an Inverse Burr distribution, and
shape1 == 1 and shape2 == shape3 is an Inverse Paralogistic distribution.

Value

dtrbeta gives the density, ptrbeta gives the distribution function, qtrbeta gives the quantile function, and rtrbeta generates random deviates.
Invalid arguments will return warning.

Note

Also known as the generalized beta of the second kind and Pearson Type VI.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dtrbeta(2, 2, 3, 4, 5, log = TRUE))
p <- (1:10)/10
ptrbeta(qtrbeta(p, 2, 3, 4, 5), 2, 3, 4, 5)
qpearson6(0.3, 2, 3, 4, 5, lower.tail = FALSE)

[Package actuar version 0.1-5 Index]