Trgamma(actuar)	R Documentation

The Transformed Gamma Distribution

Description

Density, distribution function, quantile function and random generation for the Transformed Gamma distribution with parameters shape1, scale and shape2.

Usage


  dtrgamma(x, shape1, scale, shape2, log = FALSE)
  ptrgamma(q, shape1, scale, shape2, lower.tail = TRUE,
           log.p = FALSE)
  qtrgamma(p, shape1, scale, shape2, lower.tail = TRUE,
           log.p = FALSE)
  rtrgamma(n, shape1, scale, shape2)

Arguments

x, q 	 vector of quantiles.

p 	 vector of probabilities.

n 	 number of observations. If length(n) > 1, the length is taken to be the number required.

shape1, scale, shape2 	 parameters. Must be strictly positive.

log, log.p 	 logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail 	 logical; if TRUE (default), probabilities are P[X <= x], otherwise, Pr[X > x].

Details

The Transformed Gamma distribution with parameters shape1 = a, scale = b and shape2 = c has density:

p(x) = c (x / b)^(a c) exp(-(x / b)^c) / (x gamma(a))

for x > 0, a > 0, b > 0 and c > 0. Here Gamma(a) is the function implemented by R's gamma() and defined in its help.

Let Y be a Gamma random variable with parameters shape1 and scale.

Y^(1/shape2)

has the Transformed Gamma probability density function with parameters shape1, scale and shape2.

The Transformed Gamma probability distribution defines a family of distributions with the following special cases:

shape2 == 1 is a Gamma distribution,
shape2 == shape1 == 1 is an Exponential distribution, and
shape1 == 1 is a Weibull distribution.

Value

dtrgamma gives the density, ptrgamma gives the distribution function, qtrgamma gives the quantile function, and rtrgamma generates random deviates.
Invalid arguments will return warning.

Note

Also known as the generalized gamma distribution.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Appendix A of Klugman, Panjer & Willmot, Loss Models, Second Edition, Wiley, 2004.

Examples


exp(dtrgamma(2, 3, 4, 5, log = TRUE))
p <- (1:10)/10
ptrgamma(qtrgamma(p, 2, 3, 4), 2, 3, 4)

[Package actuar version 0.1-5 Index]