Paralogistic {actuar}	R Documentation

The Paralogistic Distribution

Description

Density function, distribution function, quantile function, random generation, raw moments and limited moments for the Paralogistic distribution with parameters shape and scale.

Usage

dparalogis(x, shape, rate = 1, scale = 1/rate, log = FALSE)
pparalogis(q, shape, rate = 1, scale = 1/rate,
           lower.tail = TRUE, log.p = FALSE)
qparalogis(p, shape, rate = 1, scale = 1/rate,
           lower.tail = TRUE, log.p = FALSE)
rparalogis(n, shape, rate = 1, scale = 1/rate)
mparalogis(order, shape, rate = 1, scale = 1/rate)
levparalogis(limit, shape, rate = 1, scale = 1/rate,
             order = 1)

Arguments

x, q
vector of quantiles.

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

shape, scale
parameters. Must be strictly positive.

rate
an alternative way to specify the scale.

log, log.p
logical; if TRUE, probabilities/densities p are returned as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x].

order
order of the moment.

limit
limit of the loss variable.

Details

The paralogistic distribution with parameters shape = a and scale = s has density:

f(x) = a^2 (x/s)^a / (x [1 + (x/s)^a]^(a + 1))

for x > 0, a > 0 and b > 0.

The kth raw moment of the random variable X is E[X^k], -shape < k < shape^2.

The kth limited moment at some limit d is E[min(X, d)^k], k > -shape and shape - k/shape not a negative integer.

Value

dparalogis gives the density, pparalogis gives the distribution function, qparalogis gives the quantile function, rparalogis generates random deviates, mparalogis gives the kth raw moment, and levparalogis gives the kth moment of the limited loss variable.

Invalid arguments will result in return value NaN, with a warning.

Note

levparalogis computes the limited expected value using betaint.

See Kleiber and Kotz (2003) for alternative names and parametrizations.

Author(s)

Vincent Goulet vincent.goulet@act.ulaval.ca and Mathieu Pigeon

References

Kleiber, C. and Kotz, S. (2003), Statistical Size Distributions in Economics and Actuarial Sciences, Wiley.

Klugman, S. A., Panjer, H. H. and Willmot, G. E. (2012), Loss Models, From Data to Decisions, Fourth Edition, Wiley.

Examples

exp(dparalogis(2, 3, 4, log = TRUE))
p <- (1:10)/10
pparalogis(qparalogis(p, 2, 3), 2, 3)

## variance
mparalogis(2, 2, 3) - mparalogis(1, 2, 3)^2

## case with shape - order/shape > 0
levparalogis(10, 2, 3, order = 2)

## case with shape - order/shape < 0
levparalogis(10, 1.25, 3, order = 2)

[Package actuar version 2.3-2 Index]