actuar-deprecated {actuar}	R Documentation

Deprecated Functions in Package actuar

Description

These functions are provided for compatibility with older versions of actuar only, and may be defunct as soon as the next release.

Usage

# Deprecated in \pkg{actuar} 2.0-0
minvGauss(order, nu, lambda)
levinvGauss(limit, nu, lambda, order = 1)
mgfinvGauss(x, nu, lambda, log= FALSE)

Arguments

order
order of the moment. Only order = 1 is supported by levinvGauss.

limit
limit of the loss variable.

nu, lambda
parameters. Must be strictly positive.

x
numeric vector.

log
logical; if TRUE, the cumulant generating function is returned.

Details

Functions [m,lev,mgf]invGauss complemented functions [dpqr]invGauss of package SuppDists. From version 2.0-0, actuar has its own complete set of support functions for the inverse Gaussian distribution; see dinvgauss.

[Package actuar version 2.3-3 Index]