hachemeister(actuar)	R Documentation

Hachemeister data set

Description

Hachemeister (1975) data set giving average claim amounts in automobile liability insurance in five U.S. states between July 1970 and June 1973 and the corresponding number of claims.

Usage

data(hachemeister)

Format

A list with two components:

claims

a 5 times 12 matrix of average claim amounts

weights

a 5 times 12 matrix of corresponding weights

Source

Hachemeister, C. A. (1975), Credibility for regression models with application to trend, Proceedings of the Berkeley Actuarial Research Conference on Credibility, Academic Press.

[Package actuar version 0.1-5 Index]